requirements.txt
# For environment variables
python-dotenv

# Backend and Database Packages
flask
flask-mysqldb

# Auth
PyJWT